backend/requirements.txt
bcrypt
Flask
kiteconnect
twilio
Flask-PyMongo
flask-cors
Flask-JWT-Extended
pymongo
pycryptodome
